chore(header/store): move batch to a separate file as it deserves
